refactor(timekeeper): move ...timekeeper.utils classes to ...timekeeper
