+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/timekeeper/EventConstants.java
@@ -9,8 +9,23 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
+package com.blockchaintp.sawtooth.timekeeper;
 
 /**
- * Contains common library utility methods and classes for sawtooth-timekeeper.
+ * Constants relating to Event subscription and attributes.
  */
-package com.blockchaintp.sawtooth.timekeeper.util;
+public final class EventConstants {
+  /**
+   * Events of type timeKeeper will have this attribute set.
+   */
+  public static final String TIMEKEEPER_MICROS_ATTRIBUTE = "micros";
+
+  /**
+   * Event type for DamlLogEntry events.
+   */
+  public static final String TIMEKEEPER_EVENT_SUBJECT = Namespace.TIMEKEEPER_FAMILY_NAME + "/timeKeeper";
+
+  private EventConstants() {
+
+  }
+}
