+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/timekeeper/Namespace.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.sawtooth.timekeeper.util;
+package com.blockchaintp.sawtooth.timekeeper;
 
 import com.blockchaintp.utils.SawtoothClientUtils;
