+++ timekeeper/src/main/java/com/blockchaintp/sawtooth/timekeeper/TimeKeeperRunnable.java
@@ -22,7 +22,6 @@ import org.slf4j.LoggerFactory;
 
 import com.blockchaintp.sawtooth.timekeeper.exceptions.TimeKeeperException;
 import com.blockchaintp.sawtooth.timekeeper.protobuf.TimeKeeperUpdate;
-import com.blockchaintp.sawtooth.timekeeper.util.Namespace;
 import com.blockchaintp.utils.KeyManager;
 import com.blockchaintp.utils.SawtoothClientUtils;
 import com.google.protobuf.ByteString;
